requirements.txt
numpy
scipy
matplotlib
pandas
seaborn
torch
torchvision
torchaudio
gymnasium
stable-baselines3
jupyter
rl_zoo3
sbx-rl
tensorboard
pyyaml
pygame
Pillow
numba
opencv-python
torchviz